sql: minor improvements around estimated row count
This commit makes it so that we use the ceiling of the estimated row
count in scans and aggregation. It also fixes recently-introduced bug
that could cause integer overflow in the agg alloc code.

Release note: None

+++ pkg/sql/opt/exec/factory.go
@@ -66,7 +66,9 @@ type ScanParams struct {
 	// Row-level locking properties.
 	Locking opt.Locking
 
-	EstimatedRowCount float64
+	// EstimatedRowCount, if set, is the estimated number of rows that will be
+	// scanned, rounded up.
+	EstimatedRowCount uint64
 
 	// If true, we are performing a locality optimized search. In order for this
 	// to work correctly, the execution engine must create a local DistSQL plan
